+++ frontend/src/stores/optimisticScheduledEvents.ts
